backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.109.2
google-api-core==2.17.1
google-auth==2.28.0
google-cloud-kms==2.21.1
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httptools==0.6.1
idna==3.6
proto-plus==1.23.0
protobuf==4.25.2
pure-protobuf==3.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
